+++ CHANGELOG.md
@@ -12,7 +12,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
-- The `jsonschema` processor no longer lowercases fields within error messages.
+- The `json_schema` processor no longer lowercases fields within error messages.
 - The `dynamodb` cache no longer creates warning logs for get misses.
 
 ## 3.20.0 - 2020-07-05